+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This changelog is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [unreleased] (Breaking 🐱 | Feat 🚀 | Fixed 🐞)
 
+### Fixed 🐞
+
+- Remove prerelease suffix when bumping prerelease instead of incrementing version.
+
 ## [1.1.0] - 2026-01-11 ([full diff](https://github.com/shaman-apprentice/release-parrot/compare/v1.0.0...v1.1.0))
 
 ### Feat 🚀